Add type definition uint256

+++ packages/smart-contracts/src/contracts/ERC20EscrowToPayV1.sol
@@ -18,7 +18,7 @@ contract ERC20EscrowToPay {
         IERC20 tokenAddress;
         address payee;
         address payer;
-        uint amount;
+        uint256 amount;
         bool isFrozen;
         uint unlockDate;
     }
